- Refactored forms to be normalized - Refactored settings to be modular - Needs success state re-implemented - Needs optimization
