+++ package.json
@@ -49,7 +49,7 @@
     "@vitest/browser": "3.0.8",
     "@vitest/coverage-v8": "3.0.8",
     "autoprefixer": "^10.4.20",
-    "bits-ui": "^1.3.10",
+    "bits-ui": "^1.3.19",
     "clsx": "^2.1.1",
     "eslint": "^9.18.0",
     "eslint-config-prettier": "^10.0.1",
