+++ src/lib/components/ui/tabs/tabs-list.svelte
@@ -1,0 +1,15 @@
+<script lang="ts">
+  import { Tabs as TabsPrimitive } from 'bits-ui';
+  import { cn } from '$lib/utils.js';
+
+  let { ref = $bindable(null), class: className, ...restProps }: TabsPrimitive.ListProps = $props();
+</script>
+
+<TabsPrimitive.List
+  bind:ref
+  class={cn(
+    'inline-flex h-10 items-center justify-center rounded-md bg-muted p-1 text-muted-foreground',
+    className,
+  )}
+  {...restProps}
+/>
