+++ src/lib/components/ui/textarea/index.ts
@@ -1,7 +1,7 @@
 import Root from './textarea.svelte';
 
 export {
-	Root,
-	//
-	Root as Textarea,
+  Root,
+  //
+  Root as Textarea,
 };
