+++ src/lib/components/ui/textarea/textarea.svelte
@@ -1,22 +1,22 @@
 <script lang="ts">
-	import type { WithElementRef, WithoutChildren } from 'bits-ui';
-	import type { HTMLTextareaAttributes } from 'svelte/elements';
-	import { cn } from '$lib/utils.js';
+  import type { WithElementRef, WithoutChildren } from 'bits-ui';
+  import type { HTMLTextareaAttributes } from 'svelte/elements';
+  import { cn } from '$lib/utils.js';
 
-	let {
-		ref = $bindable(null),
-		value = $bindable(),
-		class: className,
-		...restProps
-	}: WithoutChildren<WithElementRef<HTMLTextareaAttributes>> = $props();
+  let {
+    ref = $bindable(null),
+    value = $bindable(),
+    class: className,
+    ...restProps
+  }: WithoutChildren<WithElementRef<HTMLTextareaAttributes>> = $props();
 </script>
 
 <textarea
-	bind:this={ref}
-	class={cn(
-		'border-input bg-background ring-offset-background placeholder:text-muted-foreground focus-visible:ring-ring flex min-h-[80px] w-full rounded-md border px-3 py-2 text-base focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50 md:text-sm',
-		className,
-	)}
-	bind:value
-	{...restProps}
+  bind:this={ref}
+  class={cn(
+    'flex min-h-[80px] w-full rounded-md border border-input bg-background px-3 py-2 text-base ring-offset-background placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50 md:text-sm',
+    className,
+  )}
+  bind:value
+  {...restProps}
 ></textarea>
