+++ src/lib/contexts/my-user-context.svelte.ts
@@ -496,7 +496,11 @@ export class MyUserContext {
     }
   }
 
-  async deleteMyUser(cause: string | undefined, description: string | undefined, deletePhysically: boolean): Promise<true | string> {
+  async deleteMyUser(
+    cause: string | undefined,
+    description: string | undefined,
+    deletePhysically: boolean,
+  ): Promise<true | string> {
     if (!this.client.isInitialized) {
       console.error('MyUserContext.deleteMyUser: not initialized.');
       return translate(AppUiMessage.systemError);
@@ -504,7 +508,11 @@ export class MyUserContext {
 
     try {
       isLoading = true;
-      const response = await this.client.operations.myUser.deleteMyUser(cause, description, deletePhysically);
+      const response = await this.client.operations.myUser.deleteMyUser(
+        cause,
+        description,
+        deletePhysically,
+      );
       if (response.error) {
         console.error('MyUserContext.deleteMyUser: received error.', { response });
         return translate(response.error, AppUiMessage.systemError);
@@ -521,7 +529,6 @@ export class MyUserContext {
       isLoading = false;
     }
   }
-
 
   public get isInitialized(): boolean {
     return this.client.isInitialized;
