+++ src/routes/settings/(data)/schema.ts
@@ -1,17 +1,13 @@
 import { z } from 'zod';
 
-const currentPasswordSchema = z
-  .string()
-  .min(1, {
-    message: 'Current password is required ',
-  })
-  .transform((val) => val.trim());
+export const currentPasswordSchema = z.string().min(8, {
+  message: 'Current password is required ',
+});
 
 export const emailSchema = z.string().email({
   message: 'Please enter a valid email address.',
 });
 
-// Username schema
 export const usernameSchema = z.object({
   username: z.string().min(3, 'Username must be at least 3 characters').max(30),
 });
@@ -40,7 +36,7 @@ export const passwordSchema = z.object({
 
 // Delete account schema
 export const deleteAccountSchema = z.object({
-  confirmEmail: z.string().email('Please enter a valid email address'),
+  confirmEmail: z.string().email('You must enter your current email address.'),
   reason: z.string().optional(),
   description: z.string().optional(),
 });
