+++ src/routes/settings/+layout.svelte
@@ -1,8 +1,11 @@
 <script lang="ts">
-  import { cn } from '$lib/utils';
-  import { Button } from '$lib/components/ui/button';
   import { page } from '$app/state';
   import { goto } from '$app/navigation';
+  import * as Tabs from '$lib/components/ui/tabs/index.js';
+  import { myUserContext } from '@/contexts/my-user-context.svelte';
+  import { onMount } from 'svelte';
+
+  const isSignedIn = $derived(myUserContext.isSignedIn);
 
   const tabs = [
     { id: 'account', label: 'Account', path: '/settings/account' },
@@ -10,40 +13,39 @@
   ];
 
   let { children } = $props();
+
   let activeTab = $derived.by(() => {
     const path = page.url.pathname;
-    // If we're at /settings, default to account tab without changing URL
     if (path === '/settings') {
       return 'account';
     }
     return tabs.find((tab) => path.startsWith(tab.path))?.id || 'account';
   });
+
+  onMount(() => {
+    // TODO: This redirection should happen earlier but we are using the client to determine auth.
+    // Maybe we can try to make it less jarring with Skeleton?
+    if (!isSignedIn) {
+      goto('/signin', { replaceState: true });
+    }
+  });
 </script>
 
-<div class="container py-6">
+<div class="container py-8">
   <h1 class="font-lexend text-3xl font-bold tracking-tight">Settings</h1>
-  <!-- Settings Navigation -->
-  <div class="mt-8 border-b">
-    <div class="flex space-x-8">
-      {#each tabs as tab}
-        <Button
-          variant="ghost"
-          class={cn(
-            'relative h-9 rounded-none border-b-2 border-transparent px-4',
-            activeTab === tab.id
-              ? 'border-primary font-medium text-foreground'
-              : 'text-muted-foreground',
-          )}
-          onclick={() => goto(tab.path)}
-        >
-          {tab.label}
-        </Button>
-      {/each}
-    </div>
-  </div>
 
-  <!-- Settings Content -->
-  <div class="mt-8">
-    {@render children()}
-  </div>
+  <Tabs.Root value={activeTab} class="my-8">
+    <Tabs.List class="mx-auto grid w-3/5 grid-cols-2 border-b">
+      {#each tabs as tab}
+        <Tabs.Trigger value={tab.id} onclick={() => goto(tab.path)}>
+          {tab.label}
+        </Tabs.Trigger>
+      {/each}
+    </Tabs.List>
+    {#each tabs as tab}
+      <Tabs.Content value={tab.id}>
+        {@render children()}
+      </Tabs.Content>
+    {/each}
+  </Tabs.Root>
 </div>
