+++ src/routes/settings/+page.ts
@@ -1,17 +1,19 @@
-import { superValidate } from 'sveltekit-superforms';
-import { zod } from 'sveltekit-superforms/adapters';
-import {
-  changeEmailschemaLastStep,
-  deleteAccountSchema,
-  passwordSchema,
-  usernameSchema,
-} from './account/schema';
+import type { PageLoad } from './$types';
 
-export const load = async () => {
-  return {
-    usernameForm: await superValidate(zod(usernameSchema)),
-    emailForm: await superValidate(zod(changeEmailschemaLastStep)),
-    passwordForm: await superValidate(zod(passwordSchema)),
-    deleteAccountForm: await superValidate(zod(deleteAccountSchema)),
-  };
+export const load: PageLoad = async ({ url }) => {
+  let currentTab = '';
+
+  if (url.pathname === '/settings') {
+    return {
+      currentTab: 'account',
+    };
+  }
+
+  if (url.pathname.includes('/account')) {
+    currentTab = 'account';
+  } else if (url.pathname.includes('/notifications')) {
+    currentTab = 'notifications';
+  }
+
+  return { currentTab };
 };
