+++ src/routes/settings/account/+page.svelte
@@ -1,10 +1,5 @@
 <script lang="ts">
-  import type { PageData } from './$types';
-  import AccountSettingsMenu from './components/account-settings-menu.svelte';
-
-  let { data }: { data: PageData } = $props();
+  import AccountSettingsMenu from './account-settings.svelte';
 </script>
 
-<div class="space-y-6">
-  <AccountSettingsMenu {data} />
-</div>
+<AccountSettingsMenu />
