--- src/routes/settings/account/+page.ts
@@ -1,5 +1,0 @@
-// Import the parent route's load function
-import { load as parentLoad } from '../+page';
-
-// Reuse the parent route's load function
-export const load = parentLoad;
