+++ src/routes/settings/notifications/+page.svelte
@@ -1,7 +1,5 @@
 <script lang="ts">
-  import NotificationSettingsMenu from './notification-settings-menu.svelte';
+  import NotificationSettingsMenu from './notification-settings.svelte';
 </script>
 
-<div class="space-y-6">
-  <NotificationSettingsMenu />
-</div>
+<NotificationSettingsMenu />
